move our files from renaissance folder
